+++ Supporting.cs
@@ -5,102 +5,121 @@ using System.Security.Cryptography.X509Certificates;
 using System.Text;
 using System.Threading.Tasks;
 
+//Tic Tac Toe Assignment
+
+// team 2-8
+// Cristian Cruz, Caitlyn Stokes, Perla Valdovinos, Asante Laryea-Akrong
+
 namespace Mission_4
 {
     internal class Supporting
     {
-        Program p = new Program();
 
-        /*public string TicTacPrintV2(char board)
-        { 
-            Console.WriteLine(board);
-        }*/
-  
-       public string ChatTicTacPrint(char[,] board)
+        //this function is what defines the coordinates for the print board function
+        public void DefineMoves(char[,] board, List<string> moves, string sign)
         {
-            Console.WriteLine("This is the Tic Tac Toe Board:");
 
-            string A = board[0, 0].ToString();
-            string B = board[0, 1].ToString();
-            string C = board[0, 2].ToString();
-            string D = board[1, 0].ToString();
-            string E = board[1, 1].ToString();
-            string F = board[1, 2].ToString();
-            string G = board[2, 0].ToString();
-            string H = board[2, 1].ToString();
-            string I = board[2, 2].ToString();
+            int latestEnteredIndex = moves.Count - 1;
 
-            string printBoard = $"{A} | {B} | {C}\n-+-+-\n{D} | {E} | {F}\n-+-+-\n{G} | {H} | {I}";
+            foreach (var move in moves)
+            {
+                if (moves.IndexOf(move) == latestEnteredIndex)
+                {
+                    {
+                        //Parse coordinates from the string
+                        int row = int.Parse(move[0].ToString());
+                        int col = int.Parse(move[3].ToString());
 
-            Console.WriteLine(printBoard);
-
-            return printBoard;
+                        if (sign == "X")
+                        {
+                            board[row, col] = 'X';
+                        }
+                        else
+                        {
+                            board[row, col] = '0';
+                        }
+                    }
+                }
+            }
         }
-       
 
-        /*public string TicTacPrint()//pass the array gathered from the driver team
+
+        // this function prints the board with the current player choices
+        public void PrintTicTacToe(char[,] board)
         {
-           // board = p.BoardPositions(); //calling from the program
-
-            Console.WriteLine(" This is the Tic Tac Toe Board:");
-
-            //1print current boad in a string
-            string A = "X";
-            string B = "0";
-            string C = "0";
-            string D = "0";
-            string E = "0";
-            string F = "0";
-            string G = "0";
-            string H = "0";
-            string I = "0";
-
-            string printBoard = $"{A} | {B} | {C}\n-+-+-\n{D} | {E} | {F}\n-+-+-\n{G} | {H} | {I}";
-
-            Console.WriteLine(printBoard);
-
-            return printBoard;
+            for (int row = 0; row < 3; row++)
+            {
+                for (int col = 0; col < 3; col++)
+                {
+                    Console.Write(board[row, col]);
+                    if (col < 2)
+                    {
+                        Console.Write(" | ");
+                    }
+                }
+                Console.WriteLine();
+                if (row < 2)
+                {
+                    Console.WriteLine("---------");
+                }
+            }
+        }
 
 
 
-            //2then change the letters to Xs and Os deppending on positions. in an if if 2,2 then replace position to x
-
-            //3return the board again and the guys will ask y to go 
-
-        } 
-        public string TicTacWin(Array board)
+        //this function checks for the winner if there is one
+        public bool CheckWinner(char[,] board)
         {
-            bool someoneWon = false;
-            string nameOfWinner = "";
-
+            bool isWinner = false;
+            string whoWon = "";
 
             for (int i = 0; i < 3; i++)
             {
+                // winner check in a row
                 if (board[i, 0] == board[i, 1] && board[i, 1] == board[i, 2] && board[i, 0] != ' ')
-                    return board[i, 0];  // Winner in a row
+                {
+                    isWinner = true;
+                    whoWon = board[i, 0].ToString();
+                }
 
-                if (board[0, i] == board[1, i] && board[1, i] == board[2, i] && board[0, i] != ' ')
-                    return board[0, i];  // Winner in a column
+                // winner check for column
+                else if (board[0, i] == board[1, i] && board[1, i] == board[2, i] && board[0, i] != ' ')
+                {
+                    isWinner = true;
+                    whoWon = board[0, i].ToString();
+
+                }
+
+                // check for diagonal
+                else if (board[0, 0] == board[1, 1] && board[1, 1] == board[2, 2] && board[0, 0] != ' ')
+                {
+                    isWinner = true;
+                    whoWon = board[0, 0].ToString();
+
+                }
+
+                else if (board[0, 2] == board[1, 1] && board[1, 1] == board[2, 0] && board[0, 2] != ' ')
+                {
+                    isWinner = true;
+                    whoWon = board[0, 2].ToString();
+
+                }
+
+            }
+            //prints who won 
+            if (isWinner == true)
+            {
+                Console.WriteLine($"The Winner is: {whoWon}!!");
+            }
+            else if (isWinner == false)
+            {
+                Console.WriteLine("\nThere is no winner, try again!");
             }
 
-           /* do
-            {
-                if ([a] (0,0) & (0,1) & (0,2))
-                {
-                    
-                    
-                }
-                else
-                {
+            //return the winner
+            return isWinner;
 
-                }
-
-            } while (!someoneWon);
-
-            return nameOfWinner
-        }*/
-        
-
-
+        }
     }
 }
+
